Update getting-started.md AppDelegate section
This change comes about as a result of this issue: https://github.com/wix/react-native-notifications/issues/662

The docs state to add `#import "RNNotifications.h"` to the top of `AppDelegate.m`, but this cause an error when building the App. The import should instead be added below the `#ifdef FB_SONARKIT_ENABLED` definition block.